+++ src/app/services/sdk/messages/team_pb.d.ts
@@ -1,4 +1,5 @@
 /* tslint:disable */
+/* tslint:disable */
 // package: msg
 // file: team.proto