+++ src/app/services/sdk/messages/team_pb.js
@@ -1,4 +1,5 @@
 /* eslint-disable */
+/* eslint-disable */
 // source: team.proto
 /**
  * @fileoverview